get_repos/requirements.txt
tqdm
python-dotenv
openai==0.28
